Fix magic number length
Previously we were using a 4-byte magic number to validate memory
content (or memory dump content), but it turns out the magic number is
only 2 bytes long. This was a problem after a full reset, in which
case the radio memory would no longer match our expected value.

Update the magic number in tmv71.TMV71 (and tmv71.TMD710), and also
improve the test_memory_restore test to verify that we're writing what
we're supposed to be writing.