Move channel import and export into the api
The primary motivation for this change was to simplify testing.
Testing import/export via the cli was difficult because click was
handling file i/o for us and we were wrapping all the API calls in a
`with <file>` context manager. Leaving the command line bits in the
cli but moving the logic into the API means we can pass an open file
to the API methods...which means our tests can just pass in an
`io.StringIO` object.

Closes #6